+++ lib/misc/crc32.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <string.h>
 #ifdef	HAVE_STDINT_H
   /* ISO C99 header, get definition of uint32_t */
   #include <stdint.h>
@@ -10,4 +11,4 @@
   typedef unsigned int	uint32_t
 #endif
 
-uint32_t compute_crc32(unsigned char *, int);
+uint32_t compute_crc32(unsigned char *, size_t);
